Move principals to arguments
