+++ Clarinet.toml
@@ -13,21 +13,6 @@ path = 'contracts/nft.clar'
 clarity_version = 2
 epoch = 2.5
 
-[contracts.airdrop]
-path = 'contracts/airdrop.clar'
-clarity_version = 2
-epoch = 2.5
-
-[contracts.airdrop2]
-path = 'contracts/airdrop2.clar'
-clarity_version = 2
-epoch = 2.5
-
-[contracts.airdrop3]
-path = 'contracts/airdrop3.clar'
-clarity_version = 2
-epoch = 2.5
-
 [repl.analysis]
 passes = ['check_checker']
