+++ README.md
@@ -1,0 +1,26 @@
+### Costs
+
+Below are the costs of running the multi-mint functionality in maximum capacity:
+
+```json
+{
+  "cost_result": {
+    "total": {
+      "write_length": 15013,
+      "write_count": 14996,
+      "read_length": 16441,
+      "read_count": 15000,
+      "runtime": 40311521
+    },
+    "limit": {
+      "write_length": 15000000,
+      "write_count": 15000,
+      "read_length": 100000000,
+      "read_count": 15000,
+      "runtime": 5000000000
+    },
+    "memory": 2474192,
+    "memory_limit": 100000000
+  }
+}
+```
