--- contracts/airdrop.clar.ejs
@@ -1,2 +1,0 @@
-(define-read-only (airdrop) (list
-<%= accounts.map(address => `'${address}`).join('\n') %>))
